requirements.txt
streamlit
tensorflow
Pillow
numpy
opencv-python
matplotlib